+++ Sources/MacGiver/MacGiverApp.swift
@@ -5,6 +5,7 @@ struct MacGiverApp: App {
     @StateObject private var appState = AppState()
     @StateObject private var batteryMonitor = BatteryMonitor()
     @StateObject private var storageMonitor = StorageMonitor()
+    @StateObject private var audioMixer = AudioMixer()
 
     var body: some Scene {
         MenuBarExtra {
@@ -12,6 +13,7 @@ struct MacGiverApp: App {
                 .environmentObject(appState)
                 .environmentObject(batteryMonitor)
                 .environmentObject(storageMonitor)
+                .environmentObject(audioMixer)
         } label: {
             Image(systemName: appState.menuBarSymbolName)
                 .symbolRenderingMode(.hierarchical)